[New Features]
Implements db folder and tests
[x] Tests passing

Implements prompting user for a version number. Defaults to 0.0.0.
Version number is written to package.json and bower.json files

Implements setting application title to name user inputs when prompted
during setup. The same name is used in the message that is logged when
the server starts up.